+++ addOns/client/CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
-
+### Fixed
+- Prevent loops when acting on elements due appended input data.
 
 ## [0.24.0] - 2026-05-07
 ### Added